test: Move tests into the same directory
This way they can share the utils and fixtures.
