+++ .dockerignore
@@ -1,5 +1,4 @@
 .venv
 cdk.out
 node_modules
-tests
 __pycache__
