+++ datalake/backend/tests/conftest.py
@@ -19,6 +19,16 @@ def batch_client():
 
 
 @pytest.fixture()
+def lambda_client():
+    return boto3.client("lambda")
+
+
+@pytest.fixture()
+def s3_client():
+    return boto3.client("s3")
+
+
+@pytest.fixture()
 def stepfunctions_client():
     return boto3.client("stepfunctions")
